Untriage functionality
Not really required but I've added it because the CSS misaligns on the
triaged page due to there being only a bin icon on the card.
I had to add a button to fix this and the button had to have *some*
kind of functionality. Untriage was the obvious feature to implement